servidor funcionando

+++ src/main/java/com/petshop/repository/ClienteRepository.java
@@ -6,4 +6,4 @@ import org.springframework.stereotype.Repository;
 import com.petshop.model.Cliente;
 
 @Repository
-public interface ClienteRepository extends JpaRepository<Cliente, Long> {}
+public interface ClienteRepository extends JpaRepository<Cliente, Integer> {}
